Add files via upload
In this lesson, Mike Albert taught me Steve Kelly about recursion using string reversal as an example. This lesson used Python as the language of choice. This lesson lasted over 2 hours, with plenty of comments, questions, and answers to back up the understanding. Without Mike's help, I would have been confused about why this worked. Mike explained this concept to me thoroughly and in a well-versed manner. Mike is a great teacher, helped solidify my understanding in not just recursion, but iteration, reversing of a string, as well as other concepts I have forgotten over the years.. I have never encountered recursion before, not even as my brief stint as a Computer Science Major at the University of New Hampshire. Mike Albert is a graduate of the Rensselaer Polytechnic Institute with a dual B.S. in Computer Engineering and Computer Science. I am proud to call Mike Albert my friend and brother.